Split client.go
