--- .nancy-ignore
@@ -1,1 +1,0 @@
-CVE-2022-32149
