+++ cilent_request.go
@@ -1,0 +1,52 @@
+package gocloak
+
+import (
+	"context"
+	"encoding/base64"
+	"github.com/go-resty/resty/v2"
+)
+
+// GetRequest returns a request for calling endpoints.
+func (g *GoCloak) GetRequest(ctx context.Context) *resty.Request {
+	var err HTTPErrorResponse
+	return injectTracingHeaders(
+		ctx, g.restyClient.R().
+			SetContext(ctx).
+			SetError(&err),
+	)
+}
+
+// GetRequestWithBearerAuthNoCache returns a JSON base request configured with an auth token and no-cache header.
+func (g *GoCloak) GetRequestWithBearerAuthNoCache(ctx context.Context, token string) *resty.Request {
+	return g.GetRequest(ctx).
+		SetAuthToken(token).
+		SetHeader("Content-Type", "application/json").
+		SetHeader("Cache-Control", "no-cache")
+}
+
+// GetRequestWithBearerAuth returns a JSON base request configured with an auth token.
+func (g *GoCloak) GetRequestWithBearerAuth(ctx context.Context, token string) *resty.Request {
+	return g.GetRequest(ctx).
+		SetAuthToken(token).
+		SetHeader("Content-Type", "application/json")
+}
+
+// GetRequestWithBearerAuthXMLHeader returns an XML base request configured with an auth token.
+func (g *GoCloak) GetRequestWithBearerAuthXMLHeader(ctx context.Context, token string) *resty.Request {
+	return g.GetRequest(ctx).
+		SetAuthToken(token).
+		SetHeader("Content-Type", "application/xml;charset=UTF-8")
+}
+
+// GetRequestWithBasicAuth returns a form data base request configured with basic auth.
+func (g *GoCloak) GetRequestWithBasicAuth(ctx context.Context, clientID, clientSecret string) *resty.Request {
+	req := g.GetRequest(ctx).
+		SetHeader("Content-Type", "application/x-www-form-urlencoded")
+	// Public client doesn't require Basic Auth
+	if len(clientID) > 0 && len(clientSecret) > 0 {
+		httpBasicAuth := base64.StdEncoding.EncodeToString([]byte(clientID + ":" + clientSecret))
+		req.SetHeader("Authorization", "Basic "+httpBasicAuth)
+	}
+
+	return req
+}
