+++ client_scope_mapping.go
@@ -1,0 +1,59 @@
+package gocloak
+
+import "context"
+
+// GetClientScopesScopeMappingsRealmRolesAvailable returns realm-level roles that are available to attach to this client scope
+func (g *GoCloak) GetClientScopesScopeMappingsRealmRolesAvailable(ctx context.Context, token, realm, clientScopeID string) ([]*Role, error) {
+	const errMessage = "could not get available realm-level roles with the client-scope"
+
+	var result []*Role
+
+	resp, err := g.GetRequestWithBearerAuth(ctx, token).
+		SetResult(&result).
+		Get(g.getAdminRealmURL(realm, "client-scopes", clientScopeID, "scope-mappings", "realm", "available"))
+
+	if err := checkForError(resp, err, errMessage); err != nil {
+		return nil, err
+	}
+
+	return result, nil
+}
+
+// GetClientScopesScopeMappingsRealmRoles returns roles associated with a client-scope
+func (g *GoCloak) GetClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string) ([]*Role, error) {
+	const errMessage = "could not get realm-level roles with the client-scope"
+
+	var result []*Role
+
+	resp, err := g.GetRequestWithBearerAuth(ctx, token).
+		SetResult(&result).
+		Get(g.getAdminRealmURL(realm, "client-scopes", clientScopeID, "scope-mappings", "realm"))
+
+	if err := checkForError(resp, err, errMessage); err != nil {
+		return nil, err
+	}
+
+	return result, nil
+}
+
+// DeleteClientScopesScopeMappingsRealmRoles deletes realm-level roles from the client-scope
+func (g *GoCloak) DeleteClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string, roles []Role) error {
+	const errMessage = "could not delete realm-level roles from the client-scope"
+
+	resp, err := g.GetRequestWithBearerAuth(ctx, token).
+		SetBody(roles).
+		Delete(g.getAdminRealmURL(realm, "client-scopes", clientScopeID, "scope-mappings", "realm"))
+
+	return checkForError(resp, err, errMessage)
+}
+
+// CreateClientScopesScopeMappingsRealmRoles creates realm-level roles to the client scope
+func (g *GoCloak) CreateClientScopesScopeMappingsRealmRoles(ctx context.Context, token, realm, clientScopeID string, roles []Role) error {
+	const errMessage = "could not create realm-level roles to the client-scope"
+
+	resp, err := g.GetRequestWithBearerAuth(ctx, token).
+		SetBody(roles).
+		Post(g.getAdminRealmURL(realm, "client-scopes", clientScopeID, "scope-mappings", "realm"))
+
+	return checkForError(resp, err, errMessage)
+}
